+++ src/index.js
@@ -3,6 +3,7 @@
 var catchErrors = require('./catchErrors');
 var electronApi = require('./electronApi');
 var log = require('./log').log;
+var scopeFactory = require('./scope');
 var transportConsole = require('./transports/console');
 var transportFile = require('./transports/file');
 var transportIpc = require('./transports/ipc');
@@ -11,7 +12,14 @@ var transportRemote = require('./transports/remote');
 module.exports = create('default');
 module.exports.default = module.exports;
 
+/**
+ * @param {string} logId
+ * @return {ElectronLog.ElectronLog}
+ */
 function create(logId) {
+  /**
+   * @type {ElectronLog.ElectronLog}
+   */
   var instance = {
     catchErrors: function callCatchErrors(options) {
       var opts = Object.assign({}, {
@@ -32,6 +40,8 @@ function create(logId) {
     },
   };
 
+  instance.scope = scopeFactory(instance);
+
   instance.transports = {
     console: transportConsole(instance),
     file: transportFile(instance),
@@ -40,11 +50,11 @@ function create(logId) {
   };
 
   instance.levels.forEach(function (level) {
-    instance[level] = log.bind(null, instance, level);
+    instance[level] = log.bind(null, instance, { level: level });
     instance.functions[level] = instance[level];
   });
 
-  instance.log = log.bind(null, instance, 'info');
+  instance.log = log.bind(null, instance, { level: 'info' });
   instance.functions.log = instance.log;
 
   return instance;