+++ CHANGELOG.md
@@ -15,7 +15,7 @@
  - New default log file path:
     - **on Linux:** `~/.config/{app name}/logs/{process type}.log`
     - **on macOS:** `~/Library/Logs/{app name}/{process type}.log`
-    - **on Windows:** `%USERPROFILE%\AppData\Roaming\{app name}\{process type}.log`
+    - **on Windows:** `%USERPROFILE%\AppData\Roaming\{app name}\logs\{process type}.log`
   
   
  - the option `file.fileName` is now `main.log`, `renderer.log` or `worker.log`